--- src/verify/dto/response/create-verify-response-dto.ts
@@ -1,6 +1,0 @@
-import { IsString } from 'class-validator';
-
-export class CreateVerifyResponseDto {
-  @IsString()
-  code: string;
-}
